+++ cpp/perspective/src/include/perspective/gnode.h
@@ -143,7 +143,7 @@ public:
      * @param ptr
      */
     void _register_context(
-        const std::string& name, t_ctx_type type, std::int64_t ptr
+        const std::string& name, t_ctx_type type, std::uintptr_t ptr
     );
 
     /**
